requirements.txt
langchain>=0.2.5
langchain-community>=0.2.4
langchain-text-splitters>=0.2.1
langchain-openai>=0.1.9
langchain-huggingface>=0.1.2
langchain-google-genai>=0.1.5
google-generativeai>=0.6.0
chromadb>=0.5.3
sentence-transformers>=2.7.0
pymupdf>=1.24.2
rank-bm25>=0.2.2
python-dotenv>=1.0.1
tiktoken>=0.7.0
openai>=1.40.0
streamlit>=1.36.0